Update configure-search-for-multi-geo.md
We ran across an instance with a customer who is using Result Sources to call SPO and it is pointing to their Enterprise Search Center site template. After testing and review of the source code (against the method IsMultiGeoAutoForwardingEnabled) it looks to be by design to change the value for "EnableMultiGeoSearch" to true when using an Enterprise search center template.

The Change added above is a note on this parameter to clarify that calls against the enterprise search center template will change the default behavior.