Update prepare-for-directory-synchronization.md
Removing this recently added note about tilde character as this was not accurate. 
We are not "ignoring" the tilde character whatsoever. We synchronize the tilde character as expected, but on top of that, we also add the UPN address as a secondary smtp address which does not have a tilde char (UPNs cannot have special characters). So in reality, we are synchronizing the proxy address with a tilde char plus the UPN as a secondary smtp address - I believe this secondary address was giving the wrong impression that the tilde char was being skipped.
Additionally, there's another secondary SMTP that is automatically added based in the MailnickName@Tenant.onmicrosoft.com (aka, MOERA address). In this case the tilde character will be replaced by an underscore '_' because the MailnickName cannot contain special characters either, as explained in this page under MailnickName.