Fixing wrong PDL code of SWE to correct value.
Somehow the wrong PDL code of SWE made it into this list. I have fixed it to the correct ISO 3166-1 alpha-3 value of CHE. I also tweaked the 'coming soon' text in the eDiscovery data location column to be more descriptive for CHE/ARE/ZAF as it kept confusing customers into thinking the geos were not yet launched for Multi-Geo itself.